+++ cognite/client/data_classes/__init__.py
@@ -278,6 +278,7 @@ from cognite.client.data_classes.transformations.schema import (
 )
 from cognite.client.data_classes.user_profiles import UserProfile, UserProfileList
 from cognite.client.data_classes.workflows import (
+    CancelExecution,
     CDFTaskOutput,
     CDFTaskParameters,
     DynamicTaskOutput,
@@ -530,6 +531,7 @@ __all__ = [
     "CoordinateReferenceSystem",
     "UserProfile",
     "UserProfileList",
+    "CancelExecution",
     "WorkflowUpsert",
     "WorkflowExecution",
     "WorkflowExecutionDetailed",