+++ cognite/client/data_classes/workflows.py
@@ -1115,3 +1115,38 @@ class WorkflowIds(UserList):
 
     def dump(self, camel_case: bool = True, as_external_id: bool = False) -> list[dict[str, Any]]:
         return [workflow_id.dump(camel_case, as_external_id_key=as_external_id) for workflow_id in self.data]
+
+
+@dataclass(frozen=True)
+class CancelExecution:
+    """
+    This class represents a Workflow Version Identifier.
+
+    Args:
+        workflow_external_id (str): The external ID of the workflow.
+        version (str, optional): The version of the workflow. Defaults to None.
+    """
+
+    id: str
+    reason: str | None = None
+
+    def as_primitive(self) -> tuple[str, str | None]:
+        return self.id, self.reason
+
+    @classmethod
+    def load(cls, resource: dict, cognite_client: CogniteClient | None = None) -> CancelExecution:
+        if "id" in resource:
+            execution_id = resource["id"]
+        else:
+            raise ValueError("Invalid input to WorkflowVersionId.load")
+
+        return cls(
+            id=execution_id,
+            reason=resource.get("reason"),
+        )
+
+    def dump(self, camel_case: bool = True) -> dict[str, Any]:
+        output = {"id": self.id}
+        if self.reason:
+            output["reason"] = self.reason
+        return output